backend/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.7
Flask==2.3.2
Flask-Cors==4.0.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
PyJWT==2.8.0
PyMuPDF==1.22.5
pytz==2023.3
referencing==0.30.2
rpds-py==0.9.2
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
